requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dj-database-url==2.3.0
Django==5.1.3
django-allauth==65.2.0
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
jwt==1.3.1
packaging==24.2
paho-mqtt==1.6.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
whitenoise==6.8.2
